[BUGFIX] Cache rFALd results for the whole request
The cache of the RemoteFalDriver was meant to store all received information
of remote files when the Driver itself was marked as Singleton. A recent
bugfix had to remove the SingletonInterface from the Driver so taht the cache was
not unique between objects anymore. The obvious and quick fix is to make it static.
I think this decision is excuseable (static is bad!) because the files on the remote server
can be considered static for the whole request.