[BUGFIX] Concentrate remote FAL operations for all files in Publish Files Module selected folder
The batchPrefetchFile feature was meant to sum up remote FAL calls for all required
files in the Overview Module (more general: diffing the sides) when all identifiers were
determined instead of requesting a single FAL API call for each file. A recent change
in the RecordFactory broke this feature for the Publish Files Module and each file was
requested in its own RPC call. This bugfix pools all identifiers of files in the Publish Files Module
before requesting information from the remote.