[FEATURE] Decouple communications adapter and use a registry to reference implementations
https://projekte.in2code.de/issues/28233

SSH was always a hardcoded dependency in in2publish, but
on some servers php-ssh2 is simply not available. The old SshConnection
class was splitted into several components to ease the separation of
responsibilities (file transmittion, task execution, et cetera).

With the introduction of the AdapterRegistry it is now possible to
choose the communication layer via configuration instead of xclasses.
Additionally tests and configuration requirements are now set dependent
on the selected adapter.