[DOCS] Update requirements and limitations
Sorting of records: If a record is published from stage to live, the current value of the column "sorting" is published. This can lead to duplicate or different sorting values between stage and live. Background: If the sorting of a record within a page is changed, all sorting values in all affected records might get changed. As we want to publish only a single record of this page, we prevent the publication of all affected records.