[TASK] Use a static utility instead of a service to merge configuration
Although using a service would most probably be more appropriate here,
using a static utility class by far does not require to modify as much
code as using a service would require:

We would use dependency injection to inject the service in the
ConfigContainer and would then have to modify over 50 files where the
ConfigContainer itself is not instantiated via dependency injection or
the ObjectManager::get() method but by using GeneralUtility::makeInstance().

This is not only more risky bus may also have a negative impact on the
overall performance.