[BUGFIX] Check for already visited records not only hierachy downwards
Errors like "Duplicate primary key" should be solved with this problem.
If e.g. a file was found via content element and news record (having the
content element as irre relation) the file would have been visited once
during news publishing but the $alreadyVisited value would be reset after
the news publishing to its value it had before, so the file would not be
marked as visited anymore.