[BUGFIX] Skip post processing of sys_file records which do not exist
The underlying issue is, that CommonRepository::findByIdentifier returns a record regardless of the actual existence of rows in the database.
Changing the method to return null if there is no record would break a lot of code and can not be done without a deprecation and replacement.
This change is merely a workaround than a full bug fix.

Resolves: https://projekte.in2code.de/issues/46160