[BUGFIX] Call the RedirectsSupport DataBender before publishing a redirect
The event PublishingOfOneRecordBegan was deprecated but the replacement RecordWasPublished is different, as it is dispatched _after_ record publishing. Hence, the DataBender was invoked after the record was published. To invoke it before publishing, the event RecordWasSelectedForPublishing is used.
Also, this change makes redirects publishable without any other change, if their source_host has been published incorrectly by the bug fixed with this commit.

Resolves: https://projekte.in2code.de/issues/63166