[BUGFIX] Fix cache clear task
The task to clear caches related to files results in abnormal
termination of the program. This is caused because the DataMapper used
to clear caches expects the BE_USER to either be null or fully
authenticated.

This fix copies over the existing logic to authenticate from the other
cache clear task. It also adds a condition ensuring the authentication
logic runs only once, as a second authentication is redundant if both
tasks get executed.