[TASK] Improve cache handling
As of now, in2publish only clears the cache of all published changes.
This is not the same behaviour as used in the core. This can cause the
following problem:

 1. Create a page with a menu_menu_subpages element.
 2. Publish that page.
 3. View that page on the foreign system.
 4. Create a child page on local.
 5. Observe the menu_menu_subpages listing the child page on local.
 6. Publish the child page.
 7. Observe the menu_menu_subpages not listing the child page on
    foreign.

To view the page in the menu_menu_subpages, you need to clear frontend
caches on foreign.

This patch changes the behaviour of in2publish to mirror that of the
backend. In order to do that, internal functionality is used.