[BUGFIX] replace f:image with default img tag
f:uri.resource enables cache busting as default.
This results in a file not found exception in combination with f:image.
f:image is not necessary so we replace it with the default img tag.
This still allows the use of cache busting

Relates: https://projekte.in2code.de/issues/72153